+++ Zilon.Core/Zilon.Core/Tactics/IAttackTarget.cs
@@ -15,8 +15,11 @@ namespace Zilon.Core.Tactics
         /// <summary>
         /// Принятие урона.
         /// </summary>
-        /// <param name="value">Значение, которое нужно принять.</param>
-        void TakeDamage(float value);
+        /// <param name="value"> Величина урона, полученного объектом. </param>
+        /// <remarks>
+        /// Это окончательный урон с учетом всех модифиаторов.
+        /// </remarks>
+        void TakeDamage(int value);
 
         /// <summary>
         /// Проверка, может ли цель быть атакована.
